// src/common/redis/dto/redis-promo-brand.dto.ts
export class CreateAutoStartPromoBrandDto {
  promo_brand_id: string;
  delay: number;
}

export class DeleteAutoStartPromoBrandDto {
  promo_brand_id: string;
}

export class CreateAutoFinishPromoBrandDto {
  promo_brand_id: string;
  delay: number;
}

export class DeleteAutoFinishPromoBrandDto {
  promo_brand_id: string;
}